fix: address seagrass review feedback
